tests/wpt/web-platform-tests/html/canvas/element/path-objects/2d.path.roundrect.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.path.roundrect.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.roundrect.negative</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.beginPath();
ctx.fillStyle = '#0f0';
ctx.roundRect(0, 0, 50, 25, [10, 0, 0, 0]);
ctx.roundRect(100, 0, -50, 25, [10, 0, 0, 0]);
ctx.roundRect(0, 50, 50, -25, [10, 0, 0, 0]);
ctx.roundRect(100, 50, -50, -25, [10, 0, 0, 0]);
ctx.fill();
// All rects drawn
_assertPixel(canvas, 25,12, 0,255,0,255, "25,12", "0,255,0,255");
_assertPixel(canvas, 75,12, 0,255,0,255, "75,12", "0,255,0,255");
_assertPixel(canvas, 25,37, 0,255,0,255, "25,37", "0,255,0,255");
_assertPixel(canvas, 75,37, 0,255,0,255, "75,37", "0,255,0,255");
// Correct corners are rounded.
_assertPixel(canvas, 1,1, 255,0,0,255, "1,1", "255,0,0,255");
_assertPixel(canvas, 98,1, 255,0,0,255, "98,1", "255,0,0,255");
_assertPixel(canvas, 1,48, 255,0,0,255, "1,48", "255,0,0,255");
_assertPixel(canvas, 98,48, 255,0,0,255, "98,48", "255,0,0,255");


});
</script>
